feat: deleteUser auth

+++ backend/controllers/user/deleteUser.js
@@ -1,9 +1,17 @@
 import { user } from '../../models/User.js';
+import { checkToken } from '../../utils/checkToken.js';
 
 export async function deleteUser(req, res) {
     const { username } = req.params;
+    const token = req.headers.cookie.split('=')[1];
 
     try {
+        if(!checkToken(token)) {
+            return res.status(401).json({
+                "message": "You must be logged in to perform this operation."
+            });
+        }
+
         await user.destroy({
             where: { username }
         });
